Try4?

+++ include/link.h
@@ -239,7 +239,7 @@ struct BlockRequest
 
 extern struct Link gLink;
 extern u16 ALIGNED(4) gRecvCmds[MAX_RFU_PLAYERS][CMD_LENGTH];
-extern u32 gBlockSendBuffer[BLOCK_BUFFER_SIZE];
+extern u16 gBlockSendBuffer[BLOCK_BUFFER_SIZE];
 extern u16 gLinkType;
 extern u32 gLinkStatus;
 extern u16 gBlockRecvBuffer[MAX_RFU_PLAYERS][BLOCK_BUFFER_SIZE / 2];
